+++ VisualStatistics/DateTimeTicker.h
@@ -12,7 +12,7 @@ public:
 
     void setDisplayUtc(bool displayUtc);
     void setOffsetFromUtc(int offset);
-    void setDateTimeVector(QVector<qint64> &&dtv);
+    void setDateTimeVector(QVector<double> &&dtv);
 
 signals:
     void skippedTicksChanged(int skipped);
@@ -29,7 +29,7 @@ private:
     int mOffsetFromUtc;
     QString mDateTimeFmt;
     QCPAxis *mParentAxis;
-    QVector<qint64> mDateTimeVector;
+    QVector<double> mDateTimeVector;
 };
 
 #endif // DATETIMETICKER_H
